backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.7.0
psutil==5.9.8
django-cloudinary-storage==0.3.0
setuptools==70.0.0